fix: repair committed mojibake corruption in comment lines across six crates
Multi-byte characters (x, left-right arrow, en/em dashes) in doc/comment
lines were double/multi-encoded between 41d9485 and today, ballooning host
lib.rs to 5.8 MB with up to 3.2 MB on one line and corrupting comment
intent. Repaired byte-level, preserving each line's terminator:

- forgotten-host: lib.rs L1270/L1279/L4622/L6861/L8346, frames.rs L664
- forgotten-core: lib.rs L1260/L3408, world_trade.rs L144
- forgotten-persistence: player_storage.rs L9
- forgotten-protocol: lib.rs L2527

Also removed orphaned doc blocks stranded in lib.rs by the bank/vips/
guilds/moderation extractions, restoring them onto their moved functions,
and fixed four pre-existing clippy lints in core/persistence/protocol
that surfaced under the current toolchain (unused re-export, empty line
after doc comment, redundant cast, unnecessary +1 <= comparison).

+++ crates/forgotten-core/src/world_vitals.rs
@@ -175,8 +175,7 @@ impl WorldState {
         if let Some(window) = self.player_food_windows.get_mut(&player_id) {
             if window.until_tick > self.tick {
                 window.elapsed_seconds = window.elapsed_seconds.saturating_add(elapsed_seconds);
-                let intervals =
-                    (window.elapsed_seconds / FOOD_REGENERATION_INTERVAL_SECONDS) as u16;
+                let intervals = window.elapsed_seconds / FOOD_REGENERATION_INTERVAL_SECONDS;
                 if intervals > 0 {
                     window.elapsed_seconds %= FOOD_REGENERATION_INTERVAL_SECONDS;
                     food_gained = (FOOD_REGENERATION_HEALTH_PER_INTERVAL)
